move : SellPriceToQuantityResponseDto 패키지 이동 :: Sell -> Product
