+++ src/main/java/bc1/gream/domain/product/dto/response/SellPriceToQuantityResponseDto.java
@@ -1,4 +1,4 @@
-package bc1.gream.domain.sell.dto.response;
+package bc1.gream.domain.product.dto.response;
 
 import lombok.Builder;
